apps/ui: split session-view into isolated component folders
Previously one 518-line `index.tsx` + 377-line stylesheet mixed
transcript rendering, input, thinking indicator, and top-level glue.
Split into three component folders (conversation, composer,
thinking-indicator), each with its own stylesheet, plus a glue layer
(`session-view/index.tsx` + `style.module.css`).

Each component is now position-agnostic: max-width centering, outer
paddings, and column alignment live in the glue. The components
describe only their intrinsic structure (internal layout, typography,
colors, animations). This makes each one independently composable
and testable, and the glue layer owns all cross-component layout
concerns in one place.

- `conversation/` — `Conversation` + `eventsToRenderItems` + the row
  primitives (`UserTurn`, `AssistantText`, `ToolUseRow`,
  `AgentQuestion`) + `findLatestProgressMessage`. Renders the
  `ThinkingIndicator` at the end as an internal child.
- `composer/` — `Composer` textarea + footer. Fragment-level root,
  no outer padding; glue wraps it in `.composerOuter` +
  `.column` for positioning.
- `thinking-indicator/` — pulsing dot, rotating label, elapsed
  counter, optional progress sub-line. `min-height` reserved so
  mounting/unmounting doesn't shift the surrounding layout.
- `session-view/style.module.css` — now owns the shared `.column`
  (760px max-width centering), `.conversationSpacing`
  (padding-top/bottom inside the scroll), `.composerOuter`
  (horizontal + bottom padding around the composer), plus header
  and scroll-container rules.

No visual changes.
