+++ apps/ui/src/components/session-view/conversation/style.module.css
@@ -1,0 +1,160 @@
+.root {
+	display: flex;
+	flex-direction: column;
+	gap: var(--wpds-dimension-padding-lg);
+}
+
+/* User bubbles act as turn anchors; give them generous breathing room above
+   and below so the user's prompt stands apart from the assistant response. */
+.userTurn {
+	display: flex;
+	flex-direction: column;
+	padding: var(--wpds-dimension-padding-xl) 0;
+}
+
+.userTurn:first-child {
+	padding-top: 0;
+}
+
+.userText {
+	align-self: flex-start;
+	padding: var(--wpds-dimension-padding-md) var(--wpds-dimension-padding-lg);
+	background-color: color-mix(in srgb, var(--wpds-color-fg-content-accent, #2563eb) 10%, transparent);
+	/* Bottom-left tight so the bubble reads as originating from the speaker
+	   label above; other corners round generously. */
+	border-radius: 18px 18px 18px var(--wpds-border-radius-md, 8px);
+	max-width: 100%;
+	font-size: var(--wpds-font-size-md);
+	line-height: 1.55;
+	color: var(--wpds-color-fg-content-neutral);
+	white-space: pre-wrap;
+	word-wrap: break-word;
+}
+
+.toolBlock {
+	display: flex;
+	flex-direction: column;
+	gap: 4px;
+}
+
+.toolRow {
+	display: flex;
+	align-items: baseline;
+	gap: var(--wpds-dimension-padding-sm);
+	font-family: var(--wpds-font-family-mono, ui-monospace, SFMono-Regular, Menlo, monospace);
+	font-size: var(--wpds-font-size-sm);
+	color: var(--wpds-color-fg-content-neutral-weak);
+	padding: 2px 0;
+	min-width: 0;
+}
+
+.toolOutputWrap {
+	margin-left: var(--wpds-dimension-padding-md);
+	display: flex;
+	flex-direction: column;
+	gap: 4px;
+}
+
+.toolOutput {
+	margin: 0;
+	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-md);
+	border-left: 2px solid var(--wpds-color-stroke-surface-neutral);
+	background-color: var(--wpds-color-bg-surface-neutral-strong);
+	color: var(--wpds-color-fg-content-neutral);
+	font-family: var(--wpds-font-family-mono, ui-monospace, SFMono-Regular, Menlo, monospace);
+	font-size: var(--wpds-font-size-xs);
+	line-height: 1.5;
+	white-space: pre-wrap;
+	word-wrap: break-word;
+	overflow: hidden;
+}
+
+.toolOutputCollapsed {
+	max-height: calc(1.5em * 12);
+	mask-image: linear-gradient(to bottom, black 60%, transparent 100%);
+}
+
+.toolOutputError {
+	border-left-color: var(--wpds-color-fg-content-error, #dc2626);
+	color: var(--wpds-color-fg-content-error, #dc2626);
+}
+
+.toolOutputToggle {
+	align-self: flex-start;
+	background: none;
+	border: none;
+	padding: 0;
+	color: var(--wpds-color-fg-content-accent, #2563eb);
+	font: inherit;
+	font-size: var(--wpds-font-size-xs);
+	cursor: pointer;
+}
+
+.toolOutputToggle:hover,
+.toolOutputToggle:focus-visible {
+	text-decoration: underline;
+}
+
+.toolLabel {
+	color: var(--wpds-color-fg-content-neutral-weak);
+	opacity: 0.7;
+	flex-shrink: 0;
+}
+
+.toolDetail {
+	color: var(--wpds-color-fg-content-neutral);
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+	min-width: 0;
+	flex: 1;
+}
+
+.question {
+	display: flex;
+	flex-direction: column;
+	gap: var(--wpds-dimension-padding-sm);
+}
+
+.questionText {
+	margin: 0;
+	color: var(--wpds-color-fg-content-neutral);
+}
+
+.questionOptions {
+	list-style: none;
+	margin: 0;
+	padding: 0;
+	display: flex;
+	flex-wrap: wrap;
+	gap: var(--wpds-dimension-padding-md);
+}
+
+.questionOption {
+	background: none;
+	border: none;
+	padding: 0;
+	color: var(--wpds-color-fg-content-accent, #2563eb);
+	font: inherit;
+	cursor: pointer;
+}
+
+.questionOption:hover:not(:disabled),
+.questionOption:focus-visible:not(:disabled) {
+	text-decoration: underline;
+}
+
+.questionOption:disabled {
+	color: var(--wpds-color-fg-content-neutral-weak);
+	opacity: 0.6;
+	cursor: default;
+}
+
+/* Keep the picked option fully legible even when disabled, so the user
+   can still see their choice once the batch is finalized. */
+.questionOptionPicked,
+.questionOptionPicked:disabled {
+	font-weight: 600;
+	color: var(--wpds-color-fg-content-accent, #2563eb);
+	opacity: 1;
+}
